For test using enable_user_defined_functions:true add experimental:true

+++ run.py
@@ -60,47 +60,49 @@ class Run:
 
         cmd = ["bash", "-c", "mvn  -B install -DskipTests=true -Dmaven.javadoc.skip=true -V"]
         logging.info(cmd)
-        subprocess.call(cmd, env=self._environment(), cwd=self._java_driver_git)
+        self.status = subprocess.run(cmd, env=self._environment(), cwd=self._java_driver_git)
+
+        if not self.status.returncode == 0:
+            logging.error("Build failed")
+            return
 
         if self._tag.startswith('3.7'):
             self._report_path =  "driver-core/target/surefire-reports/"
             cmd = ["bash", "-c", "rm -rf {0}*".format(self._report_path)]
             logging.info(cmd)
-            subprocess.call(cmd, env=self._environment(), cwd=self._java_driver_git)
+            subprocess.run(cmd, env=self._environment(), cwd=self._java_driver_git)
             if  self._scylla_version:
-                cmd = ["bash",  "-c",  "mvn -B -pl driver-core -Dtest.groups='long' -Dtest='{0}*' test -Dscylla.version={1}".format(exclude_str, self._scylla_version)]
+                cmd = ["bash",  "-c",  "mvn -B -pl driver-core -Dtest.groups='long' -Dtest='{0}{self._tests}' test -Dscylla.version={self._scylla_version}".format(exclude_str, self=self)]
             elif self._scylla_install_dir:
-                cmd = ["bash",  "-c",  "mvn -B -pl driver-core -Dtest.groups='long' -Dtest='{0}*' test -Dccm.directory={1}".format(exclude_str, self._scylla_install_dir)]
+                cmd = ["bash",  "-c",  "mvn -B -pl driver-core -Dtest.groups='long' -Dtest='{0}{self._tests}' test -Dccm.directory={self._scylla_install_dir}".format(exclude_str, self=self)]
             else:
                 raise ValueError("No scylla version or cassaandra dir defined")
 
             logging.info(cmd)
-            subprocess.call(cmd, env=self._environment(), cwd=self._java_driver_git)
+            self.status = subprocess.run(cmd, env=self._environment(), cwd=self._java_driver_git)
         else:
             self._report_path = "integration-tests/target/surefire-reports/"
             cmd = ["bash", "-c", "rm -rf {0}*".format(self._report_path)]
             logging.info(cmd)
-            subprocess.call(cmd, env=self._environment(), cwd=self._java_driver_git)
+            subprocess.run(cmd, env=self._environment(), cwd=self._java_driver_git)
 
             if  self._scylla_version:
-                cmd = ["bash",  "-c",  "mvn -B -pl integration-tests -Dtest='{0}*' test -Dscylla.version={1}".format(exclude_str, self._scylla_version)]
+                cmd = ["bash",  "-c",  "mvn -B -pl integration-tests -Dtest='{0}{self._tests}' test -Dscylla.version={self._scylla_version}".format(exclude_str, self=self)]
             elif self._scylla_install_dir:
-                cmd = ["bash",  "-c",  "mvn -B -pl integration-tests -Dtest='{0}*' test -Dccm.directory={1}".format(exclude_str, self._scylla_install_dir)]
+                cmd = ["bash",  "-c",  "mvn -B -pl integration-tests -Dtest='{0}{self._tests}' test -Dccm.directory={self._scylla_install_dir}".format(exclude_str, self=self)]
             else:
                 raise ValueError("No scylla version or cassaandra dir defined")
 
             logging.info(cmd)
-            subprocess.call(cmd, env=self._environment(), cwd=self._java_driver_git)
+            self.status = subprocess.run(cmd, env=self._environment(), cwd=self._java_driver_git)
 
     def _process_output(self):
         subprocess.call(["bash", "-c", "mkdir -p ../{0} ;"
                                        " rm ../{0}/* ;"
                                        " mv {1}*.* ../{0}/".format(self._tag, self._report_path)],
-
             universal_newlines=True, cwd=self._java_driver_git)
 
         subprocess.call(["bash", "-c", "sed -i 's/com.datastax/{0}.com.datastax/' ../{0}/*.xml".format(self._tag)],
                         universal_newlines=True, cwd=self._java_driver_git)
 
-        return subprocess.run(["bash", "-c", "grep -nr 'ERROR!\|FAILURE!'  ../{0}/*.txt".format(self._tag)],
-                universal_newlines=True, cwd=self._java_driver_git, stdout=subprocess.PIPE, stderr=subprocess.PIPE).stdout
+        return self.status.returncode == 0
